requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
googlesearch-python>=1.2.0
pathlib2>=2.3.0
